+++ public/simple.php
@@ -161,11 +161,11 @@ if (!empty($_REQUEST['submitted'])) {
         if ($savedsearches) { ?>
 			<div class="row">
 				<div class="col-xs-6 col-xs-offset-3">
-					<h5 style="color:slategray;margin-top:60px;"><i class="glyphicon glyphicon-time"></i> Search history</h5>
+					<h5 style="margin-top:60px;"><i class="glyphicon glyphicon-time"></i> Search history</h5>
 					<div class="well well-sm">
 						<?php
         foreach ($savedsearches as $key => $value) {
-            echo '<a class="small" style="color:slategray;" href=/simple.php?submitted=true&p=1&q=' . rawurlencode($value) . '>' . $value . '</a><br />';
+            echo '<a class="small" href=/simple.php?submitted=true&p=1&q=' . rawurlencode($value) . '>' . $value . '</a><br />';
         }
     } ?>
 					</div>